refactor: layout/index.js 파일을 export 배럴 패턴으로 수정 (#137)
